ID property auto generation. closes #23

+++ pet-clinic-data/src/main/java/com/atech/service/OwnerService.java
@@ -2,7 +2,7 @@ package com.atech.service;
 
 import com.atech.entity.Owner;
 
-public interface OwnerService extends CrudService<Owner, Integer> {
+public interface OwnerService extends CrudService<Owner, Long> {
 
     Owner findByLastName(String lastName);
